split out resolver
